Move FM flight path code to futbub
Modularizes the flight path illustration code by moving MissionPath and
several global values to futbub.cpp. This change allows the component
functions used for drawing flight path sections (and several globals) to
be hidden within futbub.cpp entirely.
