+++ src/game/future.h
@@ -1,7 +1,6 @@
 #ifndef FUTURE_H
 #define FUTURE_H
 
-void Bd(int x, int y);
 void Future(char plr);
 void MissionName(int val, int xx, int yy, int len);
